[FIX] idscocr: Free Zone Prices
Adding a new configuration field "Free Zone" (`free_zone`) that will be
company-specific, that will allow us to hide Odoo's Taxes and enable
the specific tax, freight, shipping, and margin per sale line.
Now the freight in the subtotal will be multiplied by the qty.
And also, set those values to the subtotal in the sale line, and in the
unit price of the invoice that is made from a sale.