[FIX] tracking_order: Traceback message without `subtype_`
When sending the 'internal' message from the `mail.message` model
there was a condition to create the tracking message where if the
subtype was different from a note the message will be created from
the subtype name and message type, but that condition failed if the
message didn't have a `subtype_id` assigned because it was trying to
obtain the name of an empty record.

Also, the message was being created 2 times without a real change.