[IMP] tracking_order: Tracking Events
Adding 3 new models to allow us to have better control in the
`tracking.event` model:
    - Tracking Event Location (`tracking.event.location`): It
    will contain locations for the user to set in the tracking
    events, so we can know where the event took place.
    They can be internal or externals (cases like Miami, NYC, etc),
    the externals will be used in the tracking message with the
    source type 'Manual', and the internals will be used in the
    tracking messages with the source type 'Odoo Internal'.
    - Tracking Event Message (`tracking.event.message`): It will
    contain default values for the messages in the tracking event
    (These messages will be set into the `message` field of the
    `tracking.event` model).
    - Tracking Event States (`tracking.event.state`): It will
    contain states for the user to set in the tracking events,
    so they can know in which state of the process the order is
    (These will be set in the tracking messages with the source
    type ''Manual').

In addition to those models, they will be 3 new fields for the
`tracking.event` model where the models are being added and a
group called "Tracking Order Manager", only the users with that group
will create/update/delete records of the models mentioned before.
Also, a new field is added into the `stock.location` model that will
contain the tracking location event so it can be set in the
tracking events of the pickings (this will be obtained from the
`location_dest_id` of the picking).

Making only one tree view and one form view for the model `tracking.event`
model in order to use them in the 3 models where that model is (`sale.order`,
`purchase.order` and `stock.picking`) to make it more sustainable.

And making the field `reference` in the `tracking.event` model editable
for the manual events, and adding it to the form and tree views.