[FIX] idscocr: Price Unit in SO Report Rounding
Rounding the price unit in the SOL of the Sale Order Report
so it uses the precision rounding of the currency in the sale.
Also, now that section on the report will only be replace it when the
company of the order is a "Free Zone" otherwise it will have its
original value.