[FIX] idscocr: Copy and Visible Customer Reference in PO
Now will be possible to copy the customer reference
(`client_order_ref`) when a purchase order is being duplicated, also
this field will always be visible instead of only showing in it when is
filled.
Also, we allow the track of this field to know when and by who it's
changed.